ci/requirements.txt
wheel
setuptools
Django
python-dotenv
GitPython
redis
celery
Pillow
psycopg2-binary
social-auth-app-django
django-crispy-forms
easy-thumbnails
django-image-cropping==1.4.0
sh
django-tinymce
python-socketio
eventlet
redis
django-modeltranslation==0.17.1
django-rosetta==0.9.6